# Conflicts:
#	ios/maps/MCTextureLoader.swift

+++ ios/graphics/Model/Polygon/PolygonGroup2d.swift
@@ -75,7 +75,7 @@ class PolygonGroup2d: BaseGraphicsObject {
 
         encoder.drawIndexedPrimitives(type: .triangle,
                                       indexCount: indicesCount,
-                                      indexType: .uint16,
+                                      indexType: .uint32,
                                       indexBuffer: indicesBuffer,
                                       indexBufferOffset: 0)
